+++ src/templates/node-express-ts/prettier.config.js
@@ -1,6 +1,7 @@
 module.exports = {
-  tabWidth: 2,
+  // tabWidth: 2,
   semi: true,
   singleQuote: true,
-  trailingComma: 'es5',
+  arrowParens: 'avoid',
+  // trailingComma: 'es5',
 };