+++ src/templates/node-express-ts/src/__tests__/first.spec.ts
@@ -1,0 +1,9 @@
+import '@config/env'
+
+import { PORT } from '@config/index'
+
+describe('First', () => {
+  it('should', () => {
+    expect("3333").toBe(PORT);
+  });
+})