+++ src/templates/node-express-ts/src/__tests__/first.test.ts
@@ -1,5 +1,0 @@
-describe('First', () => {
-  it('should', () => {
-    expect(2 + 2).toBe(4);
-  });
-})