+++ src/templates/node-express-ts/src/config/index.ts
@@ -1,4 +1,4 @@
-const port = process.env.PORT || 3333;
+const PORT = process.env.PORT || 3333;
 const baseUrl = process.env.URL;
 
-export { port, baseUrl };
+export { PORT, baseUrl };