+++ src/templates/node-express-ts/src/errors/index.ts
@@ -1,0 +1,3 @@
+export * from './HandleErrorNotFound';
+export * from './HandleMiddlewareError';
+export * from './HttpException';