+++ src/templates/node-express-ts/src/index.ts
@@ -1,7 +1,0 @@
-import '@config/env';
-import { baseUrl, port } from '@config/index';
-import server from './server';
-
-server.listen(port, () =>
-  console.log(`Server started go to ${baseUrl}:${port} ✨`)
-);