+++ src/templates/node-express-ts/src/routes/index.ts
@@ -1,9 +1,9 @@
 import { Request, Response, Router } from 'express';
 
-const routes = Router();
+const router = Router();
 
-routes.get('/', (_req: Request, res: Response) => {
+router.get('/', (_req: Request, res: Response) => {
   return res.json({ message: 'Welcome to Generator NodeJS API' });
 });
 
-export default routes;
+export { router };